schema for property energy added, energy data fetched and stored to database
