+++ ETL/properties.sql
@@ -17,3 +17,17 @@ IF OBJECT_ID('EnergyConsumption.dbo.Property', 'U') IS NULL
 		)
 	END
 GO
+
+
+--Table creation for the energy consumption table
+IF OBJECT_ID('EnergyConsumption.dbo.PropertyEnergy', 'U') IS NULL
+	BEGIN
+		CREATE TABLE  EnergyConsumption.dbo.PropertyEnergy (
+			propertyCode VARCHAR(100),
+			recordedDate DATE,
+			locationName VARCHAR(100),
+			EnergyType VARCHAR(100),
+			Value FLOAT
+		)
+	END
+GO
